% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CaluclateHeterogeneityEstimates.R
\name{CaluclateHeterogeneityEstimates}
\alias{CaluclateHeterogeneityEstimates}
\title{Calculate heterogeneity estimates - CODE and COMMENTS come from Xavier Basagaña}
\usage{
CaluclateHeterogeneityEstimates(
  data,
  outcome,
  conf_list,
  variable_list,
  strat_variable
)
}
\arguments{
\item{data}{A complete stacked dataset (mice object), including .imp = 0}

\item{outcome}{A string defining an outcome ("h_sdq_external" or "h_sdq_ternal")}

\item{conf_list}{A vector containing all confounder variable names}

\item{variable_list}{A data.frame containing a list of exposures for which regressions will be run, first column must contain exposure names}

\item{strat_variable}{A string identifying the name of the variable to stratify on}
}
\value{
A dataframe with heterogeneity estimates for all exposures of interest
}
\description{
Calculate heterogeneity estimates - CODE and COMMENTS come from Xavier Basagaña
}
