% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AsMidsXavi.R
\name{.AsMidsXavi}
\alias{.AsMidsXavi}
\title{Correction of the as.mids function integrated in the mice package}
\usage{
.AsMidsXavi(data, .imp, .id)
}
\arguments{
\item{data}{A complete stacked dataset (mice object) containing outcome and exposures of interest, covariate variables, .id, .imp,}

\item{.imp}{Position of the .imp column}

\item{.id}{Position of the .id column}
}
\value{
A modified mids object
}
\description{
Correction of the as.mids function integrated in the mice package
}
