% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateI.R
\name{.CalculateI}
\alias{.CalculateI}
\title{# Calculate heterogeneity measure (I squared) - CODE and COMMENTS come from Xavier Basagaña}
\usage{
.CalculateI(data, outcome, conf_list, selected_exposures, strat_variable)
}
\arguments{
\item{data}{A complete stacked dataset (mice object), including .imp = 0}

\item{outcome}{A string defining an outcome ("h_sdq_external" or "h_sdq_ternal")}

\item{conf_list}{A vector containing all confounder variable names}

\item{selected_exposures}{A vector containing all exposure variable names}

\item{strat_variable}{A string describing stratifying variable (normally "cohort")}
}
\value{
A list containing RES (estimates and CIs for all strata) and RESv (estimates and CIs for each strata separately)
}
\description{
# Calculate heterogeneity measure (I squared) - CODE and COMMENTS come from Xavier Basagaña
}
