% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pool2.R
\name{.ExpandVcov}
\alias{.ExpandVcov}
\title{Function to return an error if there are NAs}
\usage{
.ExpandVcov(q, u)
}
\arguments{
\item{q}{A name for fit coefficient}

\item{u}{u}
}
\value{
Error if NAs
}
\description{
Function to return an error if there are NAs
}
