% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FinalForestPlot.R
\name{.FinalForestPlot}
\alias{.FinalForestPlot}
\title{Plots forest plot for all exposures}
\usage{
.FinalForestPlot(selected_exposures, exposure_names, outcome, RES, RESv)
}
\arguments{
\item{selected_exposures}{A character vector listing names of exposure variables}

\item{exposure_names}{A character vector listing names of exposure variables that will be displayed on the plots}

\item{outcome}{A string defining an outcome ("h_sdq_external" or "h_sdq_internal")}

\item{RES}{A list of exposure-strat_variable interaction}

\item{RESv}{A list}
}
\value{
Forest plot for each exposure
}
\description{
Plots forest plot for all exposures
}
