% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pool2.R
\name{.MiceDf}
\alias{.MiceDf}
\title{Execution of rubin's rule depending on a sample size}
\usage{
.MiceDf(m, lambda, dfcom, method)
}
\arguments{
\item{m}{m}

\item{lambda}{lambda}

\item{dfcom}{dfcom}

\item{method}{method}
}
\value{
A number
}
\description{
Execution of rubin's rule depending on a sample size
}
