% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pool2.R
\name{.Modif}
\alias{.Modif}
\title{Function}
\usage{
.Modif(est = est, se = se, level)
}
\arguments{
\item{est}{est}

\item{se}{se}

\item{level}{level}
}
\value{
A dataframe with estimates and CIs for each stratum
}
\description{
Function
}
