% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotMetaplus2.R
\name{.PlotMetaplus2}
\alias{.PlotMetaplus2}
\title{Plot heterogeneity (forest plots)}
\usage{
.PlotMetaplus2(x, ..., extrameta = NULL, mar = c(1, 1, 1, 1))
}
\arguments{
\item{x}{A dataset of mag_meta format (for all strata, e.g. All cohorts)}

\item{...}{Parameters of the metafor::forest plot (cex, margins etc.)}

\item{extrameta}{default NULL}

\item{mar}{Default margins}
}
\value{
A forest plot
}
\description{
Plot heterogeneity (forest plots)
}
