% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pool2.R
\name{.Pool2}
\alias{.Pool2}
\title{Modified pool function for results of regressions on stacked imputed dataset}
\usage{
.Pool2(object, method = "smallsample")
}
\arguments{
\item{object}{An object which is a result of a simple regression (one outcome ~ one xposure at a time) run on the mids object created using the modified as.mids function (as.mids.xavi)}

\item{method}{default = "smallsample"}
}
\value{
A pooled mipo object
}
\description{
Modified pool function for results of regressions on stacked imputed dataset
}
