% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PoolCompare2.R
\name{.PoolCompare2}
\alias{.PoolCompare2}
\title{pool_compare2 (some function of Xavi that I do not use) calculating heterogeneity for not small sample size}
\usage{
.PoolCompare2(fit1, fit0, data = NULL, method = "Wald")
}
\arguments{
\item{fit1}{fit1}

\item{fit0}{fit0}

\item{data}{data}

\item{method}{method}
}
\value{
statistic
}
\description{
pool_compare2 (some function of Xavi that I do not use) calculating heterogeneity for not small sample size
}
