% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotHeterogeneity.R
\name{PlotHeterogeneity}
\alias{PlotHeterogeneity}
\title{Draw forest heterogeneity plots - CODE and COMMENTS come from Xavier Basagaña}
\usage{
PlotHeterogeneity(
  data,
  outcome,
  conf_list,
  variable_list,
  strat_variable,
  width,
  height,
  res,
  path,
  file_name
)
}
\arguments{
\item{data}{A complete stacked dataset (mice object), including .imp = 0}

\item{outcome}{A string defining an outcome ("h_sdq_external" or "h_sdq_internal")}

\item{conf_list}{A vector containing all confounder variable names}

\item{variable_list}{A data.frame containing a list of exposures for wich regressions will be run, first column must contain exposure names}

\item{strat_variable}{A string describing stratifying variable (normally "cohort")}

\item{width}{A scalar defining width of the plot}

\item{height}{A scalar defining height of the plot}

\item{res}{A scalar defining resolution of the plot}

\item{path}{A string defining the path for saving the plot}

\item{file_name}{A string defining the file name for saving the plot}
}
\value{
A forest plot
}
\description{
Draw forest heterogeneity plots - CODE and COMMENTS come from Xavier Basagaña
}
